# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthorReplyBody": {
        "properties": {
          "content": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "AuthorReplyBody",
        "type": "object"
      },
      "AuthorReplyResponse": {
        "properties": {
          "cascaded_ticket_count": {
            "title": "Cascaded Ticket Count",
            "type": "integer"
          },
          "hub_issue_id": {
            "title": "Hub Issue Id",
            "type": "integer"
          },
          "outbox_count": {
            "title": "Outbox Count",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "hub_issue_id",
          "version",
          "cascaded_ticket_count",
          "outbox_count"
        ],
        "title": "AuthorReplyResponse",
        "type": "object"
      },
...
      "FeedbackBody": {
        "properties": {
          "note": {
            "default": "",
            "maxLength": 2000,
            "title": "Note",
            "type": "string"
          },
          "status": {
            "pattern": "^(resolved|stillbad)$",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "FeedbackBody",
        "type": "object"
      },
      "FeedbackResponse": {
        "properties": {
          "feedback_status": {
            "title": "Feedback Status",
            "type": "string"
          },
          "hub_issue_id": {
            "title": "Hub Issue Id",
            "type": "integer"
          }
        },
        "required": [
          "hub_issue_id",
          "feedback_status"
        ],
        "title": "FeedbackResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HubIssueListResponse": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HubIssueSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "has_more"
        ],
        "title": "HubIssueListResponse",
        "type": "object"
      },
...
      "NotifyReleaseBody": {
        "properties": {
          "fix_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Fix Version",
            "type": "string"
          },
          "note": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "fix_version",
          "note"
        ],
        "title": "NotifyReleaseBody",
        "type": "object"
      },
      "NotifyReleaseResponse": {
        "properties": {
          "channel_count": {
            "title": "Channel Count",
            "type": "integer"
          },
          "hub_issue_id": {
            "title": "Hub Issue Id",
            "type": "integer"
          }
        },
        "required": [
          "hub_issue_id",
          "channel_count"
        ],
        "title": "NotifyReleaseResponse",
        "type": "object"
      },
      "OwnerSplitBody": {
        "properties": {
          "subtasks": {
            "items": {
              "$ref": "#/components/schemas/OwnerSplitSubTask"
            },
            "maxItems": 20,
            "minItems": 2,
            "title": "Subtasks",
            "type": "array"
          }
        },
        "required": [
          "subtasks"
        ],
        "title": "OwnerSplitBody",
        "type": "object"
      },
      "OwnerSplitResponse": {
        "properties": {
          "hub_issue_id": {
            "title": "Hub Issue Id",
            "type": "integer"
          },
          "sub_issues": {
            "items": {
              "$ref": "#/components/schemas/OwnerSplitSubIssueOut"
            },
            "title": "Sub Issues",
            "type": "array"
          }
        },
        "required": [
          "hub_issue_id",
          "sub_issues"
        ],
        "title": "OwnerSplitResponse",
        "type": "object"
      },
...
      "RequestSupplyBody": {
        "properties": {
          "note": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "note"
        ],
        "title": "RequestSupplyBody",
        "type": "object"
      },
...
      "SelfBugBody": {
        "properties": {
          "fix_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fix Version"
          },
          "impact_versions": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Versions"
          },
          "module": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module"
          },
          "product_line_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Line Code"
          },
          "released": {
            "default": true,
            "title": "Released",
            "type": "boolean"
          },
          "title": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "SelfBugBody",
        "type": "object"
      },
      "SelfBugResponse": {
        "properties": {
          "hub_issue_id": {
            "title": "Hub Issue Id",
            "type": "integer"
          },
          "short_code": {
            "title": "Short Code",
            "type": "string"
          }
        },
        "required": [
          "hub_issue_id",
          "short_code"
        ],
        "title": "SelfBugResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ticket-hub",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/hub-issues": {
      "get": {
        "operationId": "list_hub_issues_api_hub_issues_get",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "assigned_user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assigned User Id"
            }
          },
          {
            "in": "query",
            "name": "product",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product"
            }
          },
          {
            "in": "query",
            "name": "module",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Module"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HubIssueListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Hub Issues",
        "tags": [
          "hub-issues"
        ]
      }
    },
    "/api/hub-issues/self-bug": {
      "post": {
        "description": "登记自修复 bug：无客户来源的 standalone Bug_fix hub 工单（自查徽标）。",
        "operationId": "self_bug_endpoint_api_hub_issues_self_bug_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelfBugBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelfBugResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Self Bug Endpoint",
        "tags": [
          "hub-issues"
        ]
      }
    },
    "/api/hub-issues/{hub_issue_id}": {
      "get": {
        "operationId": "get_hub_issue_api_hub_issues__hub_issue_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "hub_issue_id",
            "required": true,
            "schema": {
              "title": "Hub Issue Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HubIssueDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Hub Issue",
        "tags": [
          "hub-issues"
        ]
      }
    },
    "/api/hub-issues/{hub_issue_id}/feedback": {
      "post": {
        "description": "记录发版后客户回访结果（resolved 闭环 / stillbad 待升级）。",
        "operationId": "feedback_endpoint_api_hub_issues__hub_issue_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "hub_issue_id",
            "required": true,
            "schema": {
              "title": "Hub Issue Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feedback Endpoint",
        "tags": [
          "hub-issues"
        ]
      }
    },
    "/api/hub-issues/{hub_issue_id}/notify-release": {
      "post": {
        "description": "发版通知：文案入 outbox（每个有源关联工单一行），回访状态置 pending。",
        "operationId": "notify_release_endpoint_api_hub_issues__hub_issue_id__notify_release_post",
        "parameters": [
          {
            "in": "path",
            "name": "hub_issue_id",
            "required": true,
            "schema": {
              "title": "Hub Issue Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotifyReleaseBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotifyReleaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Notify Release Endpoint",
        "tags": [
          "hub-issues"
        ]
      }
    },
    "/api/hub-issues/{hub_issue_id}/owner-split": {
      "post": {
        "description": "按责任人拆分（ADR-0016 P4 v1 手动）：N 个子任务 → N 个 Linear 子 issue\n（parentId 挂主 issue）+ 跟踪行。每子 issue Done 由轮询自动发 x/n 进度通知。",
        "operationId": "owner_split_endpoint_api_hub_issues__hub_issue_id__owner_split_post",
        "parameters": [
          {
            "in": "path",
            "name": "hub_issue_id",
            "required": true,
            "schema": {
              "title": "Hub Issue Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerSplitBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerSplitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Owner Split Endpoint",
        "tags": [
          "hub-issues"
        ]
      }
    },
    "/api/hub-issues/{hub_issue_id}/reply": {
      "post": {
        "description": "Author/replace the Operation reply. Cascades to linked tickets'\ncached_reply and enqueues sync_outbox rows for source write-back.",
        "operationId": "author_reply_endpoint_api_hub_issues__hub_issue_id__reply_post",
        "parameters": [
          {
            "in": "path",
            "name": "hub_issue_id",
            "required": true,
            "schema": {
              "title": "Hub Issue Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorReplyBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorReplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Author Reply Endpoint",
        "tags": [
          "hub-issues"
        ]
      }
    },
    "/api/hub-issues/{hub_issue_id}/request-supply": {
      "post": {
        "description": "Ask the customer for more info (补料). Enqueues a supply sync_outbox row\nper linked sourced ticket; the KSM sender drains them into supplyKsmOrder.",
        "operationId": "request_supply_endpoint_api_hub_issues__hub_issue_id__request_supply_post",
        "parameters": [
          {
            "in": "path",
            "name": "hub_issue_id",
            "required": true,
            "schema": {
              "title": "Hub Issue Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestSupplyBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestSupplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Supply Endpoint",
        "tags": [
          "hub-issues"
        ]
      }
    },
    "/api/hub-issues/{hub_issue_id}/urge": {
      "post": {
        "description": "催办：向 Linear issue 发评论并计数（24h 频率限制）。",
        "operationId": "urge_endpoint_api_hub_issues__hub_issue_id__urge_post",
        "parameters": [
          {
            "in": "path",
            "name": "hub_issue_id",
            "required": true,
            "schema": {
              "title": "Hub Issue Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrgeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Urge Endpoint",
        "tags": [
          "hub-issues"
        ]
      }
    },
...
  }
}